[patch from Kelvin Delbarre] Software cursor mode is now supported, although currently the existing hardware cursor mode is used whenever possible. (Software mode will be used if you are running with a recent version of SDL's Quartz video driver, since a bug in SDL 1.2.11 and later prevents the hardware cursor from working properly with that driver.)
In hardware cursor mode, the hot-spot is now determined heuristically. Formerly
it could not be determined and was always (1,1), an annoyance for many cursors
other than the arrow.

In hardware cursor mode, the cursor will now be hidden when requested by the
emulated OS (such as when you are typing in a text field).

In hardware cursor mode, some cursor image formats that the code does not handle
correctly will now be rejected, causing the emulated OS to revert temporarily to
software cursor mode. Formerly you would just end up with random garbage for a
cursor. This typically happened for grayscale or color cursors; rejecting images
with rowBytes != 2 eliminates the worst cases.