SheepShaverLauncher: Use default compiler settings
The previous Mac SDK setting, 10.4, is very old and does not allow compilation
on recent versions of OS X. Same with the compiler, gcc 4.0.

Since SheepShaver doesn't purposefully target a specific SDK or compiler,
it's better to just use XCode's default, which should always be usable.
If a user wants a specific SDK, they can set it themselves.