Basilisk II: Update README to cover bridging tun interfaces.
Now that issue #154 has been fixed, when Basilisk II is configured to
use "tun" for networking, the generated tunN interface can be bridged,
enabling AppleTalk frames to be forwarded without the Linux host
needing to be able to route AppleTalk.

This commit updates the README to discuss this possibility and briefly
outline how it can be achieved.  A number of spelling corrections and
other minor clarifications are made in the same area.