linux: x11: Video Upscaling with XRender #199
Use the XRender extention to add basic x2 upscaling of the video.

On a 4K screen with HiDPI, the video window is tiny and completely
unreadable. This patch detects HiDPI screen and attempt to upscale by
x2.

To do that, I change the rendering path so the XImage (local image) is
drawn into a server XPixmap instead of the window directly (that doesn't
support scaling)

Then after the X*PutImage is/are done, I use XRenderComposite to draw
that offscreen pixmap into the window, having applied a Transform to the
pixmap so it scales up.

The code will only work with the VOSF drawing path at the minute,
however it /should/ be easy to adapt for the other rendering paths, the
other path are already compatible with drawing into an offscreen Pixmap
now (see previous patch).

Signed-off-by: Michel Pollet <buserror@gmail.com>